fix(task): Remove older unused task store code.
